smart-timetable-ai/requirements.txt
streamlit
pandas
google-api-python-client
google-auth
google-auth-oauthlib
schedule
holidays
openai
langchain
streamlit-calendar